requirements.txt
flask
pdfminer.six
python-docx
gunicorn